+++ src/main/java/bgs/formalspecificationide/Events/Event.java
@@ -1,4 +1,4 @@
-package bgs.formalspecificationide.Services;
+package bgs.formalspecificationide.Events;
 
 /**
  * Base event class.
