+++ src/main/java/bgs/formalspecificationide/Services/EventAggregatorService.java
@@ -1,5 +1,6 @@
 package bgs.formalspecificationide.Services;
 
+import bgs.formalspecificationide.Events.Event;
 import com.google.inject.Inject;
 
 import java.util.ArrayList;
